+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/ContentManagerPlugin.java
@@ -16,14 +16,8 @@
  */
 package com.wazuh.contentmanager;
 
-import com.wazuh.contentmanager.cti.console.CtiConsole;
-import com.wazuh.contentmanager.cti.console.model.Plan;
-import com.wazuh.contentmanager.cti.console.model.Product;
-import com.wazuh.contentmanager.cti.console.model.Token;
-import com.wazuh.contentmanager.cti.console.service.AuthServiceImpl;
-import com.wazuh.contentmanager.cti.console.service.PlansServiceImpl;
+import com.wazuh.contentmanager.index.ConsumersIndex;
 import com.wazuh.contentmanager.index.ContentIndex;
-import com.wazuh.contentmanager.index.ContextIndex;
 import com.wazuh.contentmanager.settings.PluginSettings;
 import com.wazuh.contentmanager.utils.Privileged;
 import com.wazuh.contentmanager.utils.SnapshotManager;
@@ -49,12 +43,20 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.Semaphore;
 import java.util.function.Supplier;
 
-/** Main class of the Content Manager Plugin */
+/**
+ * Main class of the Content Manager Plugin
+ */
 public class ContentManagerPlugin extends Plugin implements ClusterPlugin {
     private static final Logger log = LogManager.getLogger(ContentManagerPlugin.class);
-    private ContextIndex contextIndex;
+    /**
+     * Semaphore to ensure the context index creation is only triggered once.
+     */
+    private static final Semaphore indexCreationSemaphore = new Semaphore(1);
+
+    private ConsumersIndex consumersIndex;
     private ContentIndex contentIndex;
     private SnapshotManager snapshotManager;
     private ThreadPool threadPool;
@@ -63,21 +65,21 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin {
 
     @Override
     public Collection<Object> createComponents(
-            Client client,
-            ClusterService clusterService,
-            ThreadPool threadPool,
-            ResourceWatcherService resourceWatcherService,
-            ScriptService scriptService,
-            NamedXContentRegistry xContentRegistry,
-            Environment environment,
-            NodeEnvironment nodeEnvironment,
-            NamedWriteableRegistry namedWriteableRegistry,
-            IndexNameExpressionResolver indexNameExpressionResolver,
-            Supplier<RepositoriesService> repositoriesServiceSupplier) {
+        Client client,
+        ClusterService clusterService,
+        ThreadPool threadPool,
+        ResourceWatcherService resourceWatcherService,
+        ScriptService scriptService,
+        NamedXContentRegistry xContentRegistry,
+        Environment environment,
+        NodeEnvironment nodeEnvironment,
+        NamedWriteableRegistry namedWriteableRegistry,
+        IndexNameExpressionResolver indexNameExpressionResolver,
+        Supplier<RepositoriesService> repositoriesServiceSupplier) {
         PluginSettings.getInstance(environment.settings(), clusterService);
         this.threadPool = threadPool;
         this.clusterService = clusterService;
-        this.contextIndex = new ContextIndex(client);
+        this.consumersIndex = new ConsumersIndex(client);
         this.contentIndex = new ContentIndex(client);
         this.snapshotManager =
                 new SnapshotManager(environment, this.contextIndex, this.contentIndex, new Privileged());
@@ -96,17 +98,8 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin {
     public void onNodeStarted(DiscoveryNode localNode) {
         // Only cluster managers are responsible for the initialization.
         if (localNode.isClusterManagerNode()) {
-            if (this.clusterService.state().routingTable().hasIndex(ContentIndex.INDEX_NAME)) {
-                this.start();
-            }
-
-            // To be removed once we include the Job Scheduler.
-            this.clusterService.addListener(
-                    event -> {
-                        if (event.indicesCreated().contains(ContentIndex.INDEX_NAME)) {
-                            this.start();
-                        }
-                    });
+            log.info("Starting Content Manager plugin initialization");
+            this.start();
         }
 
         /*
@@ -155,12 +148,22 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin {
     private void start() {
         try {
             this.threadPool
-                    .generic()
-                    .execute(
-                            () -> {
-                                this.contextIndex.createIndex();
-                                this.snapshotManager.initialize();
-                            });
+                .generic()
+                .execute(
+                    () -> {
+                        if (indexCreationSemaphore.tryAcquire()) {
+                            try {
+                                this.consumersIndex.createIndex();
+                            } catch (Exception e) {
+                                indexCreationSemaphore.release();
+                                log.error("Failed to create {} index, due to: {}", ConsumersIndex.INDEX_NAME, e.getMessage(), e);
+                            }
+                        } else {
+                            log.debug("{} index creation already triggered", ConsumersIndex.INDEX_NAME);
+                        }
+                        // TODO: Once initialize method is adapted to the new design, uncomment the following line
+                        //this.snapshotManager.initialize();
+                    });
         } catch (Exception e) {
             // Log or handle exception
             log.error("Error initializing snapshot helper: {}", e.getMessage(), e);
@@ -170,16 +173,16 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin {
     @Override
     public List<Setting<?>> getSettings() {
         return Arrays.asList(
-                PluginSettings.CONSUMER_ID,
-                PluginSettings.CONTEXT_ID,
-                PluginSettings.CLIENT_TIMEOUT,
-                PluginSettings.CTI_API_URL,
-                PluginSettings.CTI_CLIENT_MAX_ATTEMPTS,
-                PluginSettings.CTI_CLIENT_SLEEP_TIME,
-                PluginSettings.JOB_MAX_DOCS,
-                PluginSettings.JOB_SCHEDULE,
-                PluginSettings.MAX_CHANGES,
-                PluginSettings.MAX_CONCURRENT_BULKS,
-                PluginSettings.MAX_ITEMS_PER_BULK);
+            PluginSettings.CONSUMER_ID,
+            PluginSettings.CONTEXT_ID,
+            PluginSettings.CLIENT_TIMEOUT,
+            PluginSettings.CTI_API_URL,
+            PluginSettings.CTI_CLIENT_MAX_ATTEMPTS,
+            PluginSettings.CTI_CLIENT_SLEEP_TIME,
+            PluginSettings.JOB_MAX_DOCS,
+            PluginSettings.JOB_SCHEDULE,
+            PluginSettings.MAX_CHANGES,
+            PluginSettings.MAX_CONCURRENT_BULKS,
+            PluginSettings.MAX_ITEMS_PER_BULK);
     }
 }